complete /user/asset

+++ controller/github_controller.go
@@ -70,7 +70,7 @@ func Oauth(ctx *gin.Context){
 
 	//尝试获取数据库中该user信息
 	u := model.User{}
-	u.Id = *user.ID
+	u.Id = uint32(*user.ID)
 	u.Name = *user.Login
 	if user.AvatarURL != nil{
 		u.AvatarUrl = *user.AvatarURL
